+++ src/main/java/com/alcatrazescapee/oreveins/vein/VeinTypeSphere.java
@@ -19,7 +19,7 @@ import com.alcatrazescapee.oreveins.api.DefaultVeinType;
 public class VeinTypeSphere extends DefaultVeinType
 {
     @Override
-    public double getChanceToGenerate(DefaultVein vein, BlockPos pos)
+    public float getChanceToGenerate(DefaultVein vein, BlockPos pos)
     {
         final double dx = Math.pow(vein.getPos().getX() - pos.getX(), 2);
         final double dy = Math.pow(vein.getPos().getY() - pos.getY(), 2);
@@ -27,7 +27,6 @@ public class VeinTypeSphere extends DefaultVeinType
 
         final float radius = (float) ((dx + dz) / (horizontalSize * horizontalSize * vein.getSize()) +
                 dy / (verticalSize * verticalSize * vein.getSize()));
-
         return 0.005f * density * (1.0f - radius);
     }
 }
